move SubmissionComments files to new folder, update imports
