+++ components/ChallengeMaterial/ChallengeMaterial.tsx
@@ -25,7 +25,7 @@ import { Button } from '../theme/Button'
 import MdInput from '../MdInput'
 import { updateCache } from '../../helpers/updateCache'
 import { GlobalContext } from '../../helpers/globalContext'
-import { SubmissionComments } from '../SubmissionComments'
+import SubmissionComments from '../SubmissionComments'
 import SelectIteration from '../SelectIteration'
 import Error, { StatusCode } from '../Error'
 import ReviewStatus from '../ReviewStatus'
