+++ components/SubmissionComments/SubmissionComments.test.js
@@ -1,14 +1,14 @@
 import React, { useContext, useEffect } from 'react'
-import { SubmissionComments } from './SubmissionComments'
+import SubmissionComments from './SubmissionComments'
 import { fireEvent, waitFor, render, screen } from '@testing-library/react'
 import { MockedProvider } from '@apollo/client/testing'
-import GET_PREVIOUS_SUBMISSIONS from '../graphql/queries/getPreviousSubmissions'
+import GET_PREVIOUS_SUBMISSIONS from '../../graphql/queries/getPreviousSubmissions'
 import { InMemoryCache } from '@apollo/client'
-import submissionData from '../__dummy__/submission'
-import DELETE_COMMENT from '../graphql/queries/deleteComment'
-import EDIT_COMMENT from '../graphql/queries/editComment'
-import dummySessionData from '../__dummy__/sessionData'
-import { GlobalContext } from '../helpers/globalContext'
+import submissionData from '../../__dummy__/submission'
+import DELETE_COMMENT from '../../graphql/queries/deleteComment'
+import EDIT_COMMENT from '../../graphql/queries/editComment'
+import dummySessionData from '../../__dummy__/sessionData'
+import { GlobalContext } from '../../helpers/globalContext'
 import userEvent from '@testing-library/user-event'
 import '@testing-library/jest-dom'
